pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = "test_case_"
authors = ["matve <matveytulaevm.ama@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
pydantic = {extras = ["email"], version = "^2.7.0"}
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.1"
pillow = "^10.3.0"
bs4 = "^0.0.2"
psycopg2-binary = "^2.9.9"
install = "^1.3.5"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
drf-spectacular = "0.27.0"
setuptools = "^69.5.1"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
